Rights

+++ backend/src/main/java/com/example/backend/card/CardController.java
@@ -1,9 +1,11 @@
 package com.example.backend.card;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,6 +17,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.backend.deck.DeckRepository;
+import com.example.backend.security.Helper;
 
 @RestController
 @RequestMapping("/card")
@@ -28,10 +31,15 @@ public class CardController {
 
     @GetMapping("/{id}")
     public @ResponseBody ResponseEntity<CardEntity> getCard(@PathVariable long id) {
-        var c = cardRepository.findById(id);
-        if (c.isPresent()) {
-            return new ResponseEntity<>(c.get(), HttpStatus.OK);
-        } else {
+
+        try {
+            var card = cardRepository.findById(id).get();
+            if (!Helper.hasRightsForCard(card)) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            return new ResponseEntity<>(card, HttpStatus.OK);
+        } catch (Exception e) {
+            System.out.println(e);
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
@@ -39,27 +47,47 @@ public class CardController {
 
     @PostMapping("/find")
     public @ResponseBody ResponseEntity<List<CardEntity>> getCards(@RequestParam long deckid) {
-        var cards = cardRepository.findByDeckEntity(deckRepository.findById(deckid).get());
-
-        return new ResponseEntity<>(cards, HttpStatus.OK);
+        try {
+            var deck = deckRepository.findById(deckid).get();
+            if (!Helper.hasRightsForDeck(deck)) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            var cards = cardRepository.findByDeckEntity(deck);
+            return new ResponseEntity<>(cards, HttpStatus.OK);
+        } catch (Exception e) {
+            System.out.println(e);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @GetMapping(path = "/all")
     public @ResponseBody ResponseEntity<Iterable<CardEntity>> getAllCards() {
-        // This returns a JSON or XML with the users
-        return new ResponseEntity<>(cardRepository.findAll(), HttpStatus.OK);
+        var allCards = cardRepository.findAll();
+        List<CardEntity> allowedCards = new ArrayList<>();
+
+        for (var card : allCards) {
+            if (Helper.hasRightsForCard(card)) {
+                allowedCards.add(card);
+            }
+
+        }
+
+        return new ResponseEntity<>(allowedCards, HttpStatus.OK);
     }
 
     @PostMapping("/new")
     public @ResponseBody ResponseEntity<String> createCard(@RequestParam String hiddenPart,
             @RequestParam String visiblePart, @RequestParam(defaultValue = "0") int mark,
             @RequestParam long deckid) {
-
+        var deck = deckRepository.findById(deckid).get();
+        if (!Helper.hasRightsForDeck(deck)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         CardEntity card = new CardEntity();
         card.setHiddenPart(hiddenPart);
         card.setVisiblePart(visiblePart);
         card.setMark(mark);
-        card.setDeckEntity(deckRepository.findById(deckid).get());
+        card.setDeckEntity(deck);
 
         cardRepository.save(card);
         return new ResponseEntity<>(HttpStatus.OK);
@@ -67,6 +95,11 @@ public class CardController {
 
     @PostMapping("/delete")
     public @ResponseBody ResponseEntity<String> deleteCard(@RequestParam long id) {
+        var card = cardRepository.findById(id).get();
+        if (!Helper.hasRightsForCard(card)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         cardRepository.deleteById(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -79,6 +112,10 @@ public class CardController {
             @RequestParam(required = false, defaultValue = "-1") long deckid) {
 
         var card = cardRepository.findById(id).get();
+        if (!Helper.hasRightsForCard(card)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         if (hiddenPart != null) {
             card.setHiddenPart(hiddenPart);
         }
@@ -88,7 +125,7 @@ public class CardController {
         if (mark != -1) {
             card.setMark(mark);
         }
-        if (deckid != -1){
+        if (deckid != -1) {
             card.setDeckEntity(deckRepository.findById(id).get());
         }
         cardRepository.save(card);
